Allow automatic kernel config based on running config
To use this, add `kernel_autoconfig: True` to your vars.

This is a not so nice solution
It does call some scripts which are usually called by make
directly as they seem to behave differently when called like this

Calling `make localyesconfig || true` is also an ungly
workaround, but i did not find a way around
Calling this will convert all modules to static builtins
When calling, it will ask some questions about not configured
stuff, but you can't answer with defaults here
It fails with `Console input/output is redirected. Run 'make oldconfig' to update configuration.`
but somehow still *does* changes to the config.
Everything is converted to static now.
To be able to compile, i call `yes "" | make oldconfig`
which answers all questions with default values